Revert "RemoveAll does not fail if the file does not exist."
This reverts commit 9a0eb3277cd0b57e5086e25d1c03f72b323d8e0e.

+++ commands/create_vm_extension_test.go
@@ -47,8 +47,12 @@ var _ = Describe("CreateVMExtension", func() {
 	})
 
 	AfterEach(func() {
-		os.RemoveAll(configFile.Name())
-		os.RemoveAll(varsFile.Name())
+		if configFile != nil {
+			os.RemoveAll(configFile.Name())
+		}
+		if varsFile != nil {
+			os.RemoveAll(varsFile.Name())
+		}
 	})
 
 	Describe("Execute", func() {
